ai-service/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "insight-ai"
version = "0.1.0"
description = "AI-powered code understanding tool using LangChain and RAG"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Piyush Raj"}]
keywords = ["ai", "code-analysis", "rag", "langchain", "llm", "ollama"]

dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "click>=8.1.7",
    "rich>=13.9.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "tqdm>=4.67.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=0.3.0",
    "langchain-groq>=0.3.0",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.3.0"]
api = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
]
dev = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "isort>=5.13.2",
]
all = ["insight-ai[openai,api,dev]"]

[project.scripts]
insight = "insight.cli.main:cli"

[project.urls]
Homepage = "https://github.com/Piyush-08-bot/insight-ai"
Repository = "https://github.com/Piyush-08-bot/insight-ai"
Issues = "https://github.com/Piyush-08-bot/insight-ai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["insight*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100
